requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
transformers==4.36.2
torch>=2.2.0
datasets==2.15.0
accelerate==0.25.0
peft==0.7.1
bitsandbytes==0.41.3
tensorboard==2.15.1
wandb==0.16.1
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
pydantic==2.5.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0